优化：代码重构

+++ 2048/Classes/Game/Model/MCBlockItem.m
@@ -21,7 +21,7 @@
     if (self.empty) {
         return @"方块: (空)";
     }
-    return [NSString stringWithFormat:@"方块: (%lu)", self.value];
+    return [NSString stringWithFormat:@"方块: (%lu)", (unsigned long)self.value];
 }
 
 @end
